basic react-redux functions
useSelector is a sate hook itself. Don't need to use useState and useEffect.

You must pass a selector to useSelector, and select a value.

If you return the whole parameter, it returns the whole context value.
In that case, It triggers two renders, because you actually subscribe to the contenxt itself.
!. for the state change (if cuased by the reducer)
and 2. the context change.